deploy: cf5effb7566cfd99d08e6a09be83e252505f5837

+++ compiled/pg_duckdb_analytics/models/staging/stg_orders.sql
@@ -2,7 +2,7 @@ select
     o_orderkey::bigint as order_key,
     o_custkey::bigint as customer_key,
     o_orderstatus::text as status,
-    o_totalprice::double precision as total_price,
+    o_totalprice::numeric(15, 2) as total_price,
     o_orderdate::date as order_date,
     o_orderpriority::text as priority,
     o_clerk::text as clerk,
