+++ run/pg_duckdb_analytics/models/staging/stg_customers.sql
@@ -10,7 +10,7 @@
     c_address::text as address,
     c_nationkey::bigint as nation_key,
     c_phone::text as phone,
-    c_acctbal::double precision as account_balance,
+    c_acctbal::numeric(15, 2) as account_balance,
     c_mktsegment::text as market_segment
 from "analytics"."raw"."customer"
   );
